Updated action definitions with docker part to create appropriate images
Based on what search uses. Not totally sure this is how it should be, but it's something to review and test.
Also specified the latest version for the referenced actions. It's not like these automagically use a later version, based on documentation this can be a SHA, a release tag, or a branch name.