Updating dependencies (#18)
* Updating dependencies
Also switched to java 21. Not sure if this holds for the client library, might have to switch target to 17 there if we want to use it in 'older' projects, but lets go for 21 first.
* Adding dependency-check profile
* Removed java.version property
This is already defined in root pom with the exact same value, so not much point in declaring it again.